+++ CHANGELOG.md
@@ -7,6 +7,7 @@ GAMSPy 1.0.5
   - Allow printing the records of variable/equation attributes with a print(variable.attribute[domain].records) syntax.
   - Allow printing the records of a subset of a parameter with print(parameter['literal'].records) syntax.
   - Allow printing the records of a subset of a set with print(set['literal'].records) syntax.
+  - Update variable/equation attribute domains on addGamsCode call.
   - Show log file instead of listing file on solve statements with NEOS backend.
   - Improve the error messages of the thrown exceptions in case the user provide a model option at Container creation time. 
 - Testing